Remove Email, IdentityNumber, and IdentityType fields
This commit removes the `Email`, `IdentityNumber`, and `IdentityType` properties from the `Person` entity and related configurations. Updates were made to the `CreatePersonDto` and `PersonDto` classes to reflect these changes. A migration file was created to remove the corresponding database columns and update existing user data in the `AspNetUsers` table. Additionally, references to these properties were removed from the `FamilyMemberService`, `FamilyService`, and `PersonService` classes to ensure consistency across the application. This update simplifies the `Person` entity and its data handling.